==> code/01_generation.Rmd <==
---
title: "01_generation"
author: "Jack Alperstein"
date: "2024-01-22"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Random Generation

```{r}
pacman::p_load(tidyverse)
here::i_am("code/01_generation.Rmd")
```


```{r}
# Function to select a language based on given probabilities
select_language <- function(n) {
  languages <- c("Abkhaz", "Afar", "Afrikaans", "Akan", "Aklan", "Albanian", "Amharic", "Arabic", "Aramaic", "Armenian", "Assamese", "Assyrian", "Azerbaijani", "Bambara", "Bantu", "Basque", "Bengali", "Bihari", "Bislama", "Bosnian", "Breton", "Bulgarian", "Burmese", "Catalan", "Chamorro", "Chinese", "Chuukese", "Corsican", "Cree", "Creole", "Croatian", "Czech", "Danish", "Dari", "Dutch", "Efik", "Elamite", "English", "Eskimo", "Esperanto", "Estonian", "Ewe", "Faroese", "Farsi", "Fijian", "Filipino", "Finnish", "Flemish", "French", "French Creole", "Fujianese", "Fulani", "Georgian", "German", "Greek", "Gujarati", "Haitian Creole", "Hausa", "Hawaiian", "Hebrew", "Hindi", "Hindustani", "Hmong", "Hungarian", "Ibibio", "Icelandic", "Igbo", "Indonesian", "Inuktitut", "Irish", "Italian", "Jamaican Patois", "Japanese", "Kannada", "Kanuri", "Khmer", "Kongo", "Korean", "Kuanyama", "Kurdish", "Kwanyama", "Lao", "Latvian", "Lithuanian", "Luganda", "Macedonian", "Malay", "Malayalam", "Maltese", "Mandingo", "Marathi", "Marshallese", "Mien (Iu Mien)", "Mixtec", "Mon-Khmer", "Mongolian", "Nagamese", "Nahuatl", "Native American", "Nauruan", "Navajo", "Nepali", "Norwegian", "Occitan", "Other", "Panjabi", "Pashto", "Persian", "Pohnpeian", "Polish", "Portuguese", "Quechua", "Romanian", "Russian", "Sabah Bisaya", "Samoan", "Serbian", "Serbo-Croatian", "Sicilian", "Sign Language", "Sinhala", "Slovak", "Slovenian", "Somalian", "Spanish", "Swahili", "Swedish", "Syrian", "Tahitian", "Taiwanese", "Tamil", "Telugu", "Thai", "Tibetan", "Tigrinya", "Tongan", "Turkish", "Twi", "Uighur", "Ukrainian", "Urdu", "Vietnamese", "Walloon", "Wolof", "Yiddish", "Yoruba", "Zapotec")
  
  probabilities <- c(15, 1, 35, 2, 11, 85, 582, 12077, 12, 57069, 1, 16, 1, 1, 1, 13, 509, 11, 2, 15, 1, 133, 295, 29, 13, 66001, 2, 3, 1, 2, 264, 55, 45, 6, 201, 2, 6, 6648895, 4, 76, 369, 4, 23, 18400, 7, 32514, 29, 3, 1257, 51, 1, 1, 3, 792, 814, 939, 2, 1, 2, 892, 3217, 1, 179, 730, 1, 2, 84, 1025, 1, 7, 1440, 1, 5697, 4, 1, 806, 1, 40322, 1, 2, 1, 1009, 23, 40, 1, 26, 4, 48, 2, 34, 10, 2, 45, 28, 4623, 24, 1, 2, 6, 8, 206, 88, 191, 1, 16549, 1141, 84, 5898, 1, 531, 1126, 2, 936, 17184, 1, 1147, 215, 28, 4, 2019, 48, 31, 2, 33, 1152786, 40, 62, 6, 1, 42, 19, 112, 2646, 1, 32, 480, 315, 1, 1, 180, 1197, 54939, 2, 1, 81, 21, 3)
  
  sampled_languages <- sample(languages, n, replace = TRUE, prob = probabilities)
  
  return(sampled_languages)
}
```


```{r}
# Function to generate random dates within a specified year
random_dates <- function(n, year) {
  start_date <- as.Date(paste0(year, "-01-01"))
  end_date <- as.Date(paste0(year, "-12-31"))
  delta <- as.numeric(end_date - start_date)
  sample_dates <- start_date + sample(delta, n, replace = TRUE)
  return(sample_dates)
}
```


```{r}
# Define a mapping of counties to their respective zip codes
county_zip_mapping <- list(
  Los_Angeles = c(90001, 90002, 90003, 90004, 90005, 90006, 90007, 90008, 90009, 90010, 90011, 90012, 90013, 90014, 90015, 90016, 90017, 90018, 90019, 90020, 90021, 90022, 90023, 90024, 90025, 90026, 90027, 90028, 90029, 90030, 90031, 90032, 90033, 90034, 90035, 90036, 90037, 90038, 90039, 90040, 90041, 90042, 90043, 90044, 90045, 90046, 90047, 90048, 90049, 90050, 90051, 90052, 90053, 90054, 90055, 90056, 90057, 90058, 90059, 90060, 90061, 90062, 90063, 90064, 90065, 90066, 90067, 90068, 90069, 90070, 90071, 90072, 90073, 90074, 90075, 90076, 90077, 90078, 90079, 90080, 90081, 90082, 90083, 90084, 90086, 90087, 90088, 90089, 90090, 90091, 90093, 90094, 90095, 90096, 90097, 90099, 90101, 90102, 90103, 90174, 90185, 90189, 90201, 90202, 90209, 90210, 90211, 90212, 90213, 90220, 90221, 90222, 90223, 90224, 90230, 90231, 90232, 90233, 90239, 90240, 90241, 90242, 90245, 90247, 90248, 90249, 90250, 90251, 90254, 90255, 90260, 90261, 90262, 90263, 90264, 90265, 90266, 90267, 90270, 90272, 90274, 90275, 90277, 90278, 90280, 90290, 90291, 90292, 90293, 90294, 90295, 90296, 90301, 90302, 90303, 90304, 90305, 90306, 90307, 90308, 90309, 90310, 90311, 90312, 90313, 90397, 90398, 90401, 90402, 90403, 90404, 90405, 90406, 90407, 90408, 90409, 90410, 90411, 90501, 90502, 90503, 90504, 90505, 90506, 90507, 90508, 90509, 90510, 90601, 90602, 90603, 90604, 90605, 90606, 90607, 90608, 90609, 90610, 90612, 90631, 90637, 90638, 90639, 90640, 90650, 90651, 90652, 90659, 90660, 90661, 90662, 90665, 90670, 90671, 90701, 90702, 90703, 90704, 90706, 90707, 90710, 90711, 90712, 90713, 90714, 90715, 90716, 90717, 90723, 90731, 90732, 90733, 90734, 90744, 90745, 90746, 90747, 90748, 90749, 90755, 90801, 90802, 90803, 90804, 90805, 90806, 90807, 90808, 90809, 90810, 90813, 90814, 90815, 90822, 90831, 90832, 90833, 90834, 90835, 90840, 90842, 90844, 90845, 90846, 90847, 90848, 90853, 90888, 90895, 90899, 91001, 91003, 91006, 91007, 91008, 91009, 91010, 91011, 91012, 91016, 91017, 91020, 91021, 91023, 91024, 91025, 91030, 91031, 91040, 91041, 91042, 91043, 91046, 91066, 91077, 91101, 91102, 91103, 91104, 91105, 91106, 91107, 91108, 91109, 91110, 91114, 91115, 91116, 91117, 91118, 91121, 91123, 91124, 91125, 91126, 91129, 91131, 91175, 91182, 91184, 91185, 91186, 91187, 91188, 91189, 91191, 91201, 91202, 91203, 91204, 91205, 91206, 91207, 91208, 91209, 91210, 91214, 91221, 91222, 91224, 91225, 91226, 91301, 91302, 91303, 91304, 91305, 91306, 91307, 91308, 91309, 91310, 91311, 91312, 91313, 91316, 91321, 91322, 91324, 91325, 91326, 91327, 91328, 91329, 91330, 91331, 91333, 91334, 91335, 91337, 91340, 91341, 91342, 91343, 91344, 91345, 91346, 91350, 91351, 91352, 91353, 91354, 91355, 91356, 91357, 91361, 91364, 91365, 91367, 91371, 91372, 91376, 91380, 91381, 91382, 91383, 91384, 91385, 91386, 91387, 91388, 91390, 91393, 91399, 91401, 91402, 91403, 91405, 91406, 91409, 91411, 91423, 91436, 91497, 91501, 91502, 91504, 91505, 91506, 91601, 91602, 91604, 91605, 91606, 91607, 91608, 91702, 91706, 91711, 91722, 91723, 91724, 91731, 91732, 91733, 91734, 91735, 91740, 91741, 91744, 91745, 91746, 91747, 91748, 91749, 91750, 91754, 91755, 91756, 91759, 91765, 91766, 91767, 91768, 91769, 91770, 91771, 91772, 91773, 91775, 91776, 91778, 91780, 91788, 91789, 91790, 91791, 91792, 91793, 91795, 91797, 91799, 91801, 91802, 91803, 91804, 91841, 91896, 91899, 92397, 92821, 93243, 93510, 93532, 93534, 93535, 93536, 93539, 93543, 93544, 93550, 93551, 93552, 93553, 93563, 93584, 93586, 93590, 93591, 93599),
  Orange = c(90620, 90621, 90622, 90623, 90624, 90630, 90631, 90632, 90633, 90680, 90720, 90721, 90740, 90742, 90743, 92530, 92602, 92603, 92604, 92605, 92606, 92607, 92609, 92610, 92612, 92614, 92615, 92616, 92617, 92618, 92619, 92620, 92623, 92624, 92625, 92626, 92627, 92628, 92629, 92630, 92637, 92646, 92647, 92648, 92649, 92650, 92651, 92652, 92653, 92654, 92655, 92656, 92657, 92658, 92659, 92660, 92661, 92662, 92663, 92672, 92673, 92674, 92675, 92676, 92677, 92678, 92679, 92683, 92684, 92685, 92688, 92690, 92691, 92692, 92693, 92694, 92697, 92698, 92701, 92702, 92703, 92704, 92705, 92706, 92707, 92708, 92709, 92710, 92711, 92712, 92725, 92728, 92735, 92780, 92781, 92782, 92799, 92801, 92802, 92803, 92804, 92805, 92806, 92807, 92808, 92809, 92811, 92812, 92814, 92815, 92816, 92817, 92821, 92822, 92823, 92825, 92831, 92832, 92833, 92834, 92835, 92836, 92837, 92838, 92840, 92841, 92842, 92843, 92844, 92845, 92846, 92850, 92856, 92857, 92859, 92861, 92862, 92863, 92864, 92865, 92866, 92867, 92868, 92869, 92870, 92871, 92883, 92885, 92886, 92887, 92899),
  Ventura = c(90265, 91304, 91307, 91319, 91320, 91358, 91359, 91360, 91361, 91362, 91363, 91377, 91381, 93001, 93002, 93003, 93004, 93005, 93006, 93007, 93009, 93010, 93011, 93012, 93013, 93015, 93016, 93020, 93021, 93022, 93023, 93024, 93030, 93031, 93032, 93033, 93034, 93035, 93036, 93040, 93041, 93042, 93043, 93044, 93060, 93061, 93062, 93063, 93064, 93065, 93066, 93093, 93094, 93099, 93225, 93252),
  Riverside = c(91752, 92201, 92202, 92203, 92210, 92211, 92220, 92223, 92225, 92226, 92230, 92234, 92235, 92236, 92239, 92240, 92241, 92247, 92248, 92253, 92254, 92255, 92257, 92258, 92260, 92261, 92262, 92263, 92264, 92270, 92274, 92276, 92277, 92280, 92282, 92284, 92292, 92320, 92324, 92373, 92501, 92502, 92503, 92504, 92505, 92506, 92507, 92508, 92509, 92513, 92514, 92515, 92516, 92517, 92518, 92519, 92521, 92522, 92530, 92531, 92532, 92536, 92539, 92543, 92544, 92545, 92546, 92548, 92549, 92551, 92552, 92553, 92554, 92555, 92556, 92557, 92561, 92562, 92563, 92564, 92567, 92570, 92571, 92572, 92581, 92582, 92583, 92584, 92585, 92586, 92587, 92589, 92590, 92591, 92592, 92593, 92595, 92596, 92599, 92675, 92860, 92877, 92878, 92879, 92880, 92881, 92882, 92883),
  San_Bernardino = c(86406, 89019, 91701, 91708, 91709, 91710, 91729, 91730, 91737, 91739, 91743, 91758, 91759, 91761, 91762, 91763, 91764, 91784, 91785, 91786, 91798, 92242, 92252, 92256, 92267, 92268, 92277, 92278, 92284, 92285, 92286, 92301, 92304, 92305, 92307, 92308, 92309, 92310, 92311, 92312, 92313, 92314, 92315, 92316, 92317, 92318, 92321, 92322, 92323, 92324, 92325, 92326, 92327, 92329, 92331, 92332, 92333, 92334, 92335, 92336, 92337, 92338, 92339, 92340, 92341, 92342, 92344, 92345, 92346, 92347, 92350, 92352, 92354, 92356, 92357, 92358, 92359, 92363, 92364, 92365, 92366, 92368, 92369, 92371, 92372, 92373, 92374, 92375, 92376, 92377, 92378, 92382, 92385, 92386, 92391, 92392, 92393, 92394, 92395, 92397, 92398, 92399, 92401, 92402, 92403, 92404, 92405, 92406, 92407, 92408, 92410, 92411, 92412, 92413, 92414, 92415, 92418, 92423, 92424, 92427, 92880, 92887, 93516, 93528, 93555, 93562))
```


```{r}
generate_fake_dataset <- function(year) {
  n <- sample(c(489350, 385120, 362750, 449910, 433060, 371160, 389550, 375420, 392550, 349910), 1) # Randomly choosing the number of rows
  
  # Primary outcomes
  HEAT <- sample(c(0, 1), n, replace = TRUE, prob = c(0.99, 0.01))
  FEI <- sample(c(0, 1), n, replace = TRUE, prob = c(0.20, 0.65))
  CKD <- sample(c(0, 1), n, replace = TRUE, prob = c(0.47, 0.53))
  AKI <- sample(c(0, 1), n, replace = TRUE, prob = c(0.85, 0.15))
  RENAL <- sample(c(0, 1), n, replace = TRUE, prob = c(0.05, 0.95))
  
  # Initialize secondary outcomes with zeros
  HEAT1 <- FEI1 <- CKD1 <- AKI1 <- RENAL1 <- numeric(n)
  
  for (i in 1:n) {
    # Randomly select one outcome as primary if applicable
    outcomes <- c(HEAT[i], FEI[i], CKD[i], AKI[i], RENAL[i])
    primary_outcome_index <- if (max(outcomes) == 1) sample(which(outcomes == 1), 1) else NA
    
    if (!is.na(primary_outcome_index)) {
      if (primary_outcome_index == 1) HEAT1[i] <- 1
      if (primary_outcome_index == 2) FEI1[i] <- 1
      if (primary_outcome_index == 3) CKD1[i] <- 1
      if (primary_outcome_index == 4) AKI1[i] <- 1
      if (primary_outcome_index == 5) RENAL1[i] <- 1
    }
  }
  
  # Continue with the rest of your dataset generation as before
  counties <- names(county_zip_mapping)
  dataset <- tibble(
    DATE = random_dates(n, year),
    PATCO = sample(counties, n, replace = TRUE, prob = c(0.56, 0.15, 0.12, 0.12,0.043)), # Assign a county to each record
    # For each record, select a zip code based on the assigned county
    PATZIP = map_chr(PATCO, ~as.character(sample(county_zip_mapping[[.]], 1))),
    AGE = sample(0:100, n, replace = TRUE),
    SEX = sample(c('Male', 'Female', NA), n, replace = TRUE, prob = c(0.47, 0.53, 0.05)),
    RACE = sample(c('White', 'Black or African American', 'American Indian or Alaska Native', 
                    'Asian/Pacific Islander', 'Other Race', NA), n, replace = TRUE, prob = c(0.61, 0.12, 0.003, 0.083, 0.18, 0.004)),
    ETH = sample(c('Non-Hispanic', 'Hispanic', NA), n, replace = TRUE, prob = c(0.66, 0.34, 0.001)),
    PAYOR = sample(c('Other', 'Medicaid', 'Self-Pay', 'Workers Compensation', NA), n, replace = TRUE, prob = c(0.1, 0.32, 0.15, 0.1, 0.01)),
    LANG = select_language(n),
    HEAT = HEAT,
    FEI = FEI,
    CKD = CKD,
    AKI = AKI,
    RENAL = RENAL,
    HEAT1 = HEAT1,
    FEI1 = FEI1,
    CKD1 = CKD1,
    AKI1 = AKI1,
    RENAL1 = RENAL1
  )
  
  return(dataset)
}
```


```{r}
set.seed(123) # For reproducibility

for (i in 10:19) {
  year <- 2000 + i
  fake_dataset <- generate_fake_dataset(year)
  file_path <- paste0("data/EXTRACT", i, ".csv")
  write.csv(fake_dataset, file_path, row.names = FALSE)
}
```
